+++ js/data.js
@@ -1,0 +1,61 @@
+const ANNOUNCEMENT_COUNT = 10;
+const DECIMAL_PLACES = 5;
+const LATITUDE_START = 35.65;
+const LATITUDE_END = 35.7;
+const LONGITUDE_START = 139.7;
+const LONGITUDE_END = 139.8;
+const MAX_PRICE = 1000000;
+const MAX_ROOMS_COUNT = 20;
+const MAX_USERS_COUNT = 1000;
+
+const TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const CHECKIN = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const CHECKOUT = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+export {
+  ANNOUNCEMENT_COUNT,
+  DECIMAL_PLACES,
+  LATITUDE_START,
+  LATITUDE_END,
+  LONGITUDE_START,
+  LONGITUDE_END,
+  MAX_PRICE,
+  MAX_ROOMS_COUNT,
+  MAX_USERS_COUNT,
+  TYPE,
+  CHECKIN,
+  CHECKOUT,
+  FEATURES,
+  PHOTOS
+};
